Remove outer border around hero code
At some point after refactoring, there was a weird effect we were
getting with the hero code sample, where it was nestled inside a
background surface that was a slightly different color. This was
probably when we broke the kobweb hero demo from being to the right
of the widget to below it.

Here, we clean things up so it looks like just an undecorated code
area now.