Change order in hero section
It seems like it's a bit easier to mentally follow when the examples come
AFTER what the final result looks like, rather than before. In other words,
show the goal (as a motivation) and then how we did it.

It probably makes sense to continue this throughout the rest of the site
when we start adding docs.