Fix weird color issue with link in Kobweb site hero example
The Kobweb hero example basically toggles between dark mode and light mode,
but intentionally is separated from the global color mode (since it is trying
to demo an alternating dark / light color pattern regardless of the user's
own setting).

A long time ago I updated the way to do this in Kobweb (by using Surface
to handle it for you), but I realize now I never updated this code.
Whoops!