ProjectOrganization -> ProjectStructure
Instead of being a footnote about the "components" subfolder, it now
acts as an intro section that gives users a very high level overview of
all Kobweb projects. This is nice as it let me add notes about parts of
Kobweb that were previously under-documented, such as the .kobweb folder.